// src/player-status/eligibility.enums.ts
export enum EligibilityStatus {
  ELIGIBLE = 'eligible',
  FEE_UNPAID = 'fee_unpaid',
  NO_MEDICAL = 'no_medical',
  INJURED = 'injured',
  SUSPENDED = 'suspended',
  OTHER = 'other',
  DISABLED = 'disabled',
}

export enum ImpedimentType {
  INJURY = 'injury',
  SUSPENSION = 'suspension',
  OTHER = 'other',
}

export enum EligibilityColor {
  GREEN = 'green',
  RED = 'red',
  ORANGE = 'orange',
  YELLOW = 'yellow',
}